Refactor ES framework
* do not limit facet values to integers, but use long
* distinguish integer, long and double field types
* remove unused sub-facets

+++ server/sonar-server/src/main/java/org/sonar/server/issue/InternalRubyIssueService.java
@@ -111,7 +111,7 @@ public class InternalRubyIssueService implements ServerComponent {
     return issueService.getByKey(issueKey);
   }
 
-  public Map<String, Integer> findIssueAssignees(Map<String, Object> params) {
+  public Map<String, Long> findIssueAssignees(Map<String, Object> params) {
     return issueService.findIssueAssignees(issueQueryService.createFromMap(params));
   }
